compress paths, tested in real @rootfs

+++ spec/support/filesystem.rb
@@ -5,20 +5,28 @@ module FileSystem
   end
 
   def create_file(path)
+    path = relativate path
     create_directory path.parent
     with_optional_fakefs do |enabled|
       if enabled
         File.open(path, 'w') do |file|
           file.write precached_files[ path.basename.to_s ] ||
-            "a non-empty file with path: '#{path}'"
+            sample_content(path)
         end
       else
-        FileUtils.copy file_factory_path/path.basename, path
+        if (file_factory_path/path.basename).exist?
+          FileUtils.copy file_factory_path/path.basename, path
+        else
+          File.open(path, 'w') do |file|
+            file.write sample_content(path)
+          end
+        end
       end
     end
   end
 
   def create_directory(path)
+    path = relativate path
     with_optional_fakefs do
       FileUtils.mkdir_p path
     end
@@ -69,6 +77,19 @@ module FileSystem
 
   def disable_fakefs_on_demand!
     @fakefs = false
+  end
+
+  def relativate(path)
+    path = Pathname.new(path) unless path.is_a?(Pathname)
+    if path.relative?
+      rootfs/path
+    else
+      path
+    end
+  end
+
+  def sample_content(path=nil)
+    "a non-empty file with path: '#{path}'"
   end
 end
